refactor: canvas 위젯 이동 및 임시 placeholder 정리
