+++ lib/features/canvas/widgets/canvas_actions.dart
@@ -3,19 +3,7 @@ import 'dart:convert';
 import 'package:flutter/material.dart';
 import 'package:scribble/scribble.dart';
 
-import '../../data/sketches.dart';
-
 class CanvasActions {
-  static void saveSketch(BuildContext context, ScribbleNotifier notifier) {
-    final json = notifier.currentSketch.toJson();
-    final data = SketchData(
-      name: 'temp',
-      description: 'temp',
-      jsonData: jsonEncode(json),
-    );
-    sketches.add(data);
-  }
-
   static void showImage(BuildContext context, ScribbleNotifier notifier) async {
     final image = notifier.renderImage();
     if (!context.mounted) return;
