+++ lib/features/canvas/widgets/canvas_info.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-/// 📊 캔버스와 뷰포트 정보를 표시하는 위젯
+/// 캔버스와 뷰포트 정보를 표시하는 위젯
 class CanvasInfo extends StatelessWidget {
   const CanvasInfo({
     required this.canvasWidth,
